pyproject.toml
[project]
name = "database"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "sqladmin[full]>=0.21.0",
    "sqlalchemy-sqlitecloud>=0.1.2",
    "uvicorn>=0.35.0",
]
